% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create_templates}
\alias{create_templates}
\title{CREATE MPLUS TEMPLATES ACCORDING TO THE SPECIFICATIONS MADE}
\usage{
create_templates(lca_settings)
}
\value{
List of character vectors with model templates. Also saves them as .txt files in current working directory.
}
\description{
CREATE MPLUS TEMPLATES ACCORDING TO THE SPECIFICATIONS MADE
}
