% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_lca.R
\name{define_lca}
\alias{define_lca}
\title{Define LCA}
\usage{
define_lca(
  frame,
  analysis_name,
  id_variable,
  weight_variable = character(),
  nclasses = 4,
  starts = 160,
  cores = 16,
  use = character(),
  nominal = character(),
  categorical = character(),
  censored_above = character(),
  censored_below = character(),
  inflated = character(),
  poisson = character(),
  negbin = character()
)
}
\arguments{
\item{frame}{This dataframe contains the data you want to perform the LCA on.}

\item{analysis_name}{A name for the analysis. Please choose a meaningful short name that will also be used for constructing subfolders needed for the files. Date and time will be added to the name automatically.}

\item{id_variable}{Variable in data frame that contains the unique ids for samples. Will not be included in the analysis.}

\item{weight_variable}{Variable that contains weightings for all samples in the data set (optional).}

\item{nclasses}{The maximum number of classes the LCA should be performed for.}

\item{starts}{Number of random starts for each class and model type. It is advisable to use numbers of the form 2^X*10. Must be provided either as single integer or in the form of a List with six entries for the 6 model types, each entry comprising an integer vector of length nclasses.}

\item{cores}{The number of cores to use when performing the LCA.}

\item{use}{Character vector with all variables in the data frame used for the LCA. All other variables will be saved as auxiliary variables in the LCA and not used for computing.}

\item{nominal}{Character vector with all multinominal (unordered) variables. Multinominal variables must be stored as integers > 0.}

\item{categorical}{Character vector with all ordinal variables. Categorical variables must be stored as integers > 0.}

\item{censored_above}{Character vector with all censored above variables.}

\item{censored_below}{Character vector with all censored below variables.}

\item{inflated}{Character vector with all zero-inflated variables (can only be censored or poisson, negbin). Inflation is always assumed to vary between classes.}

\item{poisson}{Character vector with all poisson-distributed variables. Poisson-distributed variables must be positive integers.}

\item{negbin}{Character vector with all negative-binomial-distributed variables. Negative-binomial-distributed variables must be positive integers.}
}
\value{
Environment with settings for the LCA that can be passed to the \code{\link[=perform_lca]{perform_lca()}} command.
}
\description{
This command creates a settings environment for performing the easy LCA. It also performs basic checks to see if the information provided can be used to successfully perform an LCA.
}
\examples{
titanic_settings <- define_lca(
 frame = titanic_passengers,
 analysis_name = 'titanic',
 id_variable = 'id',
 nclasses = 10,
 nominal = c('port', 'pasclass'),
 categorical = c('survived', 'isfem', 'nsibsp', 'nparchi'),
 starts = 160,
 cores = 16,
)
}
