% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_data.R
\name{diagnose_data}
\alias{diagnose_data}
\title{Diagnose Data}
\usage{
diagnose_data(frame, id)
}
\arguments{
\item{frame}{The data frame you want to perform the LCA on.}

\item{id}{The id columns of the data frame. It will be excluded from the diagnosis. For all other variables a histogram will be generated and the relative frequency of NAs will be shown.}
}
\value{
List with histograms for all variables and a data frame that shows the relative frequencies of NAs in every variable at the end of the list.
}
\description{
Show histograms for all variables in the data frame (except id) and the relative frequencies of NAs for all variables in the data frame (except id). Histograms will also be drawn for categorical variables. Categorical variables must be supplied as integers with values > 0 to be interpreted the right way (as is also necessary for the LCA).
}
\examples{
diagnose_data(testdata, 'id')
}
