% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_diagnosis_report.R
\name{generate_data_diagnosis_report}
\alias{generate_data_diagnosis_report}
\title{Diagnose Data}
\usage{
generate_data_diagnosis_report(frame, idcol, title)
}
\arguments{
\item{frame}{The data frame you want to perform the LCA on.}

\item{idcol}{The id column of the data frame. It will be excluded from the diagnosis.}

\item{title}{title of the data set / project for display in the HTML document.}
}
\value{
NULL. An html file with the diagnosis will be generated in the current working directory.
}
\description{
Generates an HTML report that informs on:
\enumerate{
\item Relative frequency of Missings for all indicators
\item Correlations between continuous indicators and contigency tables between discrete indicators
\item Distributions of variables for all indicators
}
}
\details{
Discrete variables must be supplied as integers with values > 0 to be interpreted the right way (as is also necessary for the LCA). Discrete variables should not have more than 10 levels.

\strong{Coutinuous variables must be explicitly be supplied as type double}, else they will be treated as discrete variables.
}
\examples{
diagnose_data(testdata, 'id')
}
