% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_model_report.R
\name{generate_model_report}
\alias{generate_model_report}
\title{Show information on a specific model}
\usage{
generate_model_report(easylca, modeltype, classes)
}
\arguments{
\item{easylca}{easylca object as can be obtained with \code{\link[=perform_lca]{perform_lca()}} or \code{\link[=rerun_lca]{rerun_lca()}}.}

\item{modeltype}{type of model. For an explanation on model types please refer to the documentation of \code{\link[=perform_lca]{perform_lca()}}. Must have been calculated for the settings specified in the easylca object.}

\item{classes}{Which k-class solution for the modeltype to show. Must be less or equal to nclasses parameter in the easylca setting.}
}
\value{
NULL. An html file with the model information will be generated in the current working directory.
}
\description{
Generates an HTML report with detailed information on the model specified.
}
\examples{
# generate a report for model a type-2 model with 3 classes
# generate_model_report(titanic_results, 2, 3)
}
\seealso{
\code{\link[=generate_model_selection_report]{generate_model_selection_report()}}
\code{\link[=get_prediction_for_model]{get_prediction_for_model()}}
}
