% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_lca.R
\name{perform_lca}
\alias{perform_lca}
\title{Perform LCA}
\usage{
perform_lca(settings, modeltypes = seq(6))
}
\arguments{
\item{settings}{Settings for the lca, including data, variable specification, and additional technical specifications. Please use the \code{\link[=define_lca]{define_lca()}} command to generate the settings and refer to its documentation for further details.}

\item{modeltypes}{Modeltypes to perform the LCA for. Defaults to all 6 model types.}
}
\value{
easylca object that contains settings, models, and plots
}
\description{
Performs the LCA with the settings provided and saves diagnostic plots for all combinations of model types and classes as well as a summary of all models for model choice.
There are 6 modeltypes that can be estimated :
\enumerate{
\item Equal variances across classes, and covariances fixed to
\item Varying variances across classes and covariances fixed to 0
\item Equal variances and equal covariances accross classes
\item Varying variances and equal covariances across classes
\item Equal variances and varying covariances across classes
\item Varying variances and varying covariances across classes
}
}
\details{
Inspect the results with \code{\link[=generate_model_selection_report]{generate_model_selection_report()}}, \code{\link[=generate_model_report]{generate_model_report()}}, and \code{\link[=get_prediction_for_model]{get_prediction_for_model()}}.
}
\examples{
# for more model types, add them to the modeltype vector, take a while to compute
# lca <- perform_lca(titanic_settings, modeltype = c(1, 2, 3))
}
\seealso{
For a more detailed explation on the different modeltypes we recommend to visit \url{https://data-edu.github.io/tidyLPA/articles/Introduction_to_tidyLPA.html}.
}
