% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_models.R
\name{read_models}
\alias{read_models}
\title{BETA VERSION: Read Already Existing Models that match the given settings}
\usage{
read_models(settings)
}
\arguments{
\item{settings}{LCA settings object}
}
\value{
LCA object
}
\description{
Reads currently existing models according to given settings. In case something went wrong during training and a results file could not be written.
}
\details{
#TODO: Read out start values and update them in the settings object.
}
\examples{
# settings <- random_testresults$settings
# perform_lca(settings, modeltypes = c(1, 3)
# lca <- read_models(settings)
}
