% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_model.R
\name{show_model}
\alias{show_model}
\title{Show diagnosis for a specific model}
\usage{
show_model(easylca, modeltype, classes)
}
\arguments{
\item{easylca}{easylca object as can be obtained with perform_lca() or rerun_lca()}

\item{modeltype}{type of model. For an explanation on model types please refer to the documentation of perform_lca(). Must have been calculated for the settings specified in the easylca object.}

\item{classes}{Which k-class solution for the modeltype to show. Must be <= nclasses parameter in the easylca setting.}
}
\value{
Specified model as a list
}
\description{
Show diagnosis for a specific model
}
\examples{
show_model(testresults, 2, 3) # shows results for model type 2, 3 class solution
}
