% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_summary.R
\name{show_summary}
\alias{show_summary}
\title{Show LCA summary}
\usage{
show_summary(easylca)
}
\arguments{
\item{easylca}{easylca object created with \code{\link[=perform_lca]{perform_lca()}} or \code{\link[=rerun_lca]{rerun_lca()}}.}
}
\value{
Data frame with LCA summary.
}
\description{
Show LCA summary
}
\examples{
show_summary(testresults)
}
\seealso{
\code{\link[=show_model]{show_model()}}
\code{\link[=get_prediction_for_model]{get_prediction_for_model()}}
}
