% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_lca_results}
\alias{titanic_lca_results}
\title{Titanic passengers example easylca results}
\format{
An \emph{easylca} settings object with the following parameters:
\describe{
\item{settings}{settings used for the lca}
\item{models}{models for all 6 model types with up to 10 classes}
\item{summary}{results summary data frame (including VLMR test)}
}
}
\usage{
titanic_lca_results
}
\description{
An easylca object results for an LCA conducted using the \link{titanic_passengers} example data set.
}
\seealso{
\itemize{
\item \link{titanic_passengers} for the corresponding  data set
\item run \code{\link[=generate_model_selection_report]{generate_model_selection_report()}} to create a summary HTML-data report of the results
\item run \code{\link[=generate_model_report]{generate_model_report()}} to create a report on a specific model solution
}
}
\keyword{datasets}
