% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_passengers}
\alias{titanic_passengers}
\title{Titanic passengers example data set}
\format{
A data frame with 891 rows and 9 columns:
\describe{
\item{id}{Passenger ID}
\item{survived}{Has the passenger survived the sinking of the ship?}
\item{pasclass}{Passenger Class}
\item{age}{Age of the passenger}
\item{fare}{Ticket fare for the passenger}
\item{nsibsp}{Number of siblings or spouses aboard, three levels:
\describe{
\item{1}{no siblings or spouses}
\item{2}{1 sibling or spouse}
\item{3}{2 or more siblings or spouses}
}
}
\item{nparchi}{Number of parents or children aboard, three levels:
\describe{
\item{1: }{no parents or children}
\item{2: }{1 parent or children}
\item{3: }{2 or more parents or children}
}
}
\item{isfem}{Is passenger female?}
\item{port}{Port of embarkation, three levels:
\describe{
\item{1: }{Southampton}
\item{2: }{Cherbourg}
\item{3: }{Queenstown}
}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/titanic/readme/README.html}
}
\usage{
titanic_passengers
}
\description{
The training data set from the \emph{titanic} R package, prepared for an analysis with \emph{easylca}.
}
\seealso{
run \code{\link[=generate_data_diagnosis_report]{generate_data_diagnosis_report()}} to create an HTML-data report
}
\keyword{datasets}
