% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_settings}
\alias{titanic_settings}
\title{Titanic passengers example data set}
\format{
An \emph{easylca} settings object with the following parameters:
\describe{
\item{frame}{\link{titanic_passengers} data set}
\item{analysis_name}{titanic}
\item{id}{id}
\item{nclasses}{10}
\item{starts}{all start values set to 160}
\item{cores}{use 16 cores for processing}
\item{nominal}{\emph{port} and \emph{pasclass} are declared as nominal}
\item{categorical}{\emph{survived, isfem, nsibsp}, and \emph{nparchi} declared as categorical (meaning: ordered)}
\item{lmrlrt}{FALSE, no LMRLRT will be caluclated}
\item{inflation, censoring}{no inflation or censoring for the continuous variables \emph{age} and \emph{fare}}
}
}
\usage{
titanic_settings
}
\description{
A settings object with the correctly specified lca parameters for the \link{titanic_passengers} example data set.
}
\seealso{
\itemize{
\item \link{titanic_passengers} for the corresponding  data set
\item run \code{\link[=generate_data_diagnosis_report]{generate_data_diagnosis_report()}} to create an HTML-data report
}
}
\keyword{datasets}
