% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction_for_model.R
\name{get_prediction_for_model}
\alias{get_prediction_for_model}
\title{Get Model Predictions}
\usage{
get_prediction_for_model(easylca, modeltype, classes)
}
\arguments{
\item{easylca}{easylca object created with \code{\link[=perform_lca]{perform_lca()}} or \code{\link[=rerun_lca]{rerun_lca()}}.}

\item{modeltype}{type of model. For an explanation on model types please refer to the documentation of \code{\link[=perform_lca]{perform_lca()}}. Must have been calculated for the settings specified in the easylca object.}

\item{classes}{Which k-class solution for the modeltype to show. Must be less or equal to nclasses parameter in the easylca setting.}
}
\value{
Data frame including all data as well as model predictions (hard classes, probabilites for classes).
}
\description{
Get Model Predictions
}
\examples{
get_prediction_for_model(testresults, 3, 3)
}
\seealso{
\code{\link[=show_summary]{show_summary()}}
\code{\link[=show_model]{show_model()}}
}
