% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_lca.R
\name{perform_lca}
\alias{perform_lca}
\title{Perform LCA}
\usage{
perform_lca(settings, modeltypes = seq(6))
}
\arguments{
\item{settings}{for the lca, including data, variable specification, and additional technical specifications. Please use the define_lca() command to generate the settings and refer to its documentation for further details.}

\item{modeltypes}{Modeltypes to perform the LCA for. Defaults to all 6 model types.}
}
\value{
easylca object that contains settings, models, and plots
}
\description{
Performs the LCA with the settings provided and saves diagnostic plots for all combinations of model types and classes as well as a summary of all models for model choice.
}
\examples{
settings define_lca(testdata, 'test', 'id')
perform_lca(settings)
}
