% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun_lca.R
\name{rerun_lca}
\alias{rerun_lca}
\title{Rerun LCA}
\usage{
rerun_lca(easylca, models_and_starts = NULL)
}
\arguments{
\item{easylca}{easylca object obtained from the \code{\link[=perform_lca]{perform_lca()}} or \code{\link[=rerun_lca]{rerun_lca()}} command.}

\item{models_and_starts}{models to rerun with new start values. Must be a data.frame with columns: classes, modeltype, starts. If NULL (default) will rerun all models that were not replicated with doubled number of starts.}
}
\value{
easylca object with updated values
}
\description{
Rerun LCA
}
\examples{
lca <- perform_lca(testresults$settings, modeltypes = c(1,2))

# automatic update of start values and choice of models
rerun_lca(lca)

# manual update of start values and choice of models
rerun_lca(lca, data.frame(modeltype = c(1, 1), classes = c(2, 3), starts = c(30, 160)))
}
