% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_model.R
\name{show_model}
\alias{show_model}
\title{Show diagnosis for a specific model}
\usage{
show_model(easylca, modeltype, classes)
}
\arguments{
\item{easylca}{easylca object as can be obtained with \code{\link[=perform_lca]{perform_lca()}} or \code{\link[=rerun_lca]{rerun_lca()}}.}

\item{modeltype}{type of model. For an explanation on model types please refer to the documentation of \code{\link[=perform_lca]{perform_lca()}}. Must have been calculated for the settings specified in the easylca object.}

\item{classes}{Which k-class solution for the modeltype to show. Must be less or equal to nclasses parameter in the easylca setting.}
}
\value{
Specified model as a list
}
\description{
Show diagnosis for a specific model
}
\examples{
show_model(testresults, 2, 3) # shows results for model type 2, 3 class solution
}
\seealso{
\code{\link[=show_summary]{show_summary()}}
\code{\link[=get_prediction_for_model]{get_prediction_for_model()}}
}
